Fix the close method
This allows the close method to be called multiple times without causing
an error due to dangling resources. The destructor, atexit, and context
functions can now be used simultaneously without worry.

+++ py/erlang/vm.py
@@ -29,8 +29,18 @@ class VM(object):
         self.close()
 
     def close(self):
-        if self.pipe.returncode is None:
-            os.kill(self.pipe.pid, signal.SIGKILL)
+        if self.conn:
+            try:
+                self.conn.close()
+            except:
+                pass
+        if self.pipe and self.pipe.returncode is None:
+            try:
+                os.kill(self.pipe.pid, signal.SIGKILL)
+            except:
+                pass
+        self.pipe = None
+        self.conn = None
 
     def boot(self):
         sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
